+++ src/client/java/miyucomics/hexical/features/media_jar/MediaJarRenderHooks.kt
@@ -6,8 +6,8 @@ import at.petrak.hexcasting.common.lib.HexItems
 import com.mojang.datafixers.util.Pair
 import miyucomics.hexical.inits.HexicalBlocks.MEDIA_JAR_BLOCK
 import miyucomics.hexical.inits.HexicalBlocks.MEDIA_JAR_BLOCK_ENTITY
-import miyucomics.hexical.misc.DecimalFormats
 import miyucomics.hexical.misc.InitHook
+import miyucomics.hexical.misc.TextUtilities
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap
 import net.minecraft.client.render.RenderLayer
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactories
@@ -20,7 +20,7 @@ object MediaJarRenderHooks : InitHook() {
 		BlockEntityRendererFactories.register(MEDIA_JAR_BLOCK_ENTITY) { MediaJarBlockEntityRenderer() }
 		ScryingLensOverlayRegistry.addDisplayer(MEDIA_JAR_BLOCK) { lines, _, pos, _, world, _ ->
 			val jar = world.getBlockEntity(pos) as MediaJarBlockEntity
-			lines.add(Pair(ItemStack(HexItems.AMETHYST_DUST), Text.translatable("hexcasting.tooltip.media", DecimalFormats.DUST_AMOUNT.format(jar.getMedia().toFloat() / MediaConstants.DUST_UNIT.toFloat()))))
+			lines.add(Pair(ItemStack(HexItems.AMETHYST_DUST), Text.translatable("hexcasting.tooltip.media", TextUtilities.DUST_AMOUNT.format(jar.getMedia().toFloat() / MediaConstants.DUST_UNIT.toFloat()))))
 		}
 	}
 }