+++ src/main/java/miyucomics/hexical/features/prestidigitation/PrestidigitationBlockBooleans.kt
@@ -8,7 +8,6 @@ import net.fabricmc.fabric.api.resource.ResourceManagerHelper
 import net.fabricmc.fabric.api.resource.SimpleSynchronousResourceReloadListener
 import net.minecraft.block.Block
 import net.minecraft.registry.Registries
-import net.minecraft.registry.Registry
 import net.minecraft.resource.ResourceManager
 import net.minecraft.resource.ResourceType
 import net.minecraft.state.property.BooleanProperty
@@ -24,21 +23,21 @@ object PrestidigitationBlockBooleans {
 			override fun getFabricId() = HexicalMain.id("prestidigitation_block_boolean")
 			override fun reload(manager: ResourceManager) {
 				map.clear()
-				manager.findResources("prestidigitation") { it.path.endsWith("block_boolean.json") }.keys.forEach { path ->
+				manager.findResources("prestidigitation") { it.path.endsWith("block_booleans.json") }.keys.forEach { path ->
 					(JsonParser.parseReader(InputStreamReader(manager.getResource(path).get().inputStream, "UTF-8")) as JsonObject).entrySet().forEach {
-						map[Registries.BLOCK.get(Identifier(it.key))] = BooleanProperty.of(it.value.asString)
+						val id = Identifier(it.key)
+						if (Registries.BLOCK.containsId(id))
+							map[Registries.BLOCK.get(id)] = BooleanProperty.of(it.value.asString)
 					}
 				}
 			}
 		})
 
-		Registry.register(PrestidigitationHandlersHook.PRESTIDIGITATION_HANDLER, HexicalMain.id("boolean_block"), object : PrestidigitationHandler {
-			override fun tryHandleBlock(env: CastingEnvironment, position: BlockPos): Boolean {
-				val state = env.world.getBlockState(position)
-				if (state.block !in map)
-					return false
-				env.world.setBlockState(position, state.with(map[state.block], !state.get(map[state.block])))
-				return true
+		PrestidigitationHandlersHook.register("boolean_block", object : PrestidigitationHandlerBlock() {
+			override fun canAffectBlock(env: CastingEnvironment, pos: BlockPos) = map.containsKey(getBlock(env, pos))
+			override fun affect(env: CastingEnvironment, pos: BlockPos) {
+				val state = getBlockState(env, pos)
+				setBlockState(env, pos, state.with(map[state.block], !state.get(map[state.block])))
 			}
 		})
 	}